--- src/main/kotlin/io/hhplus/tdd/point/UserPoint.kt
@@ -1,7 +1,0 @@
-package io.hhplus.tdd.point
-
-data class UserPoint(
-    val id: Long,
-    val point: Long,
-    val updateMillis: Long,
-)
