+++ src/main/kotlin/io/hhplus/tdd/point/domain/UserPoint.kt
@@ -1,0 +1,7 @@
+package io.hhplus.tdd.point.domain
+
+data class UserPoint(
+    val id: Long,
+    val point: Long,
+    val updateMillis: Long,
+)
