+++ src/main/kotlin/io/hhplus/tdd/point/service/PointServiceImpl.kt
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.point.service
 
-import io.hhplus.tdd.point.PointHistory
-import io.hhplus.tdd.point.UserPoint
+import io.hhplus.tdd.point.domain.PointHistory
+import io.hhplus.tdd.point.domain.UserPoint
 import io.hhplus.tdd.point.domain.req.UserPointRequest
 import org.springframework.stereotype.Service
